[test] postman, mySQL local test : Comment Controller 중 likeCount 감소 기능 외 테스트 완료

+++ src/main/java/KWU/likelion/GoldBalance/controller/VoteController.java
@@ -15,12 +15,12 @@ public class VoteController implements VoteOperations {
     private final VoteServiceImpl voteService;
     @Override // 클라이언트가 한 투표를 받고 저장
     public ResponseEntity<String> vote(String id, String selectNum) {
-        voteService.vote((long) Integer.parseInt(id), selectNum); // 일단 id를 long으로 형 변환해서 넘기겠음
+        voteService.vote( Integer.parseInt(id), selectNum); // 일단 id를 long으로 형 변환해서 넘기겠음
         return ResponseEntity.ok("투표 완료!");
     }
     @Override    //해당 포스트에 대한 투표 결과(VoteResult 객체에) 주기 ex> 왼쪽 5표 오른쪽 8표 이런식으로
     public ResponseEntity<VoteResult> getVoteResult(String id) {
-        VoteResult voteResult = voteService.getVoteResult(Long.parseLong(id)); // 일단 id 를 long으로 바꿔서 인자 전달
+        VoteResult voteResult = voteService.getVoteResult(Integer.parseInt(id)); // 일단 id 를 long으로 바꿔서 인자 전달
         return new ResponseEntity<>(voteResult, HttpStatus.OK); // VoteResult 전달
     }
 }
